Backdrop, DetailTile components created. Rating styled components moved to Rating folder
